Add files via upload
it is our task to build a predictive model using machine learning to predict the probability
of a device failure. When building this model, be sure to minimize false positives and false
negatives. The column you are trying to predict is called failure with binary value 0 for
non-failure and 1 for failure.